+++ src/plots/scatter-plot.tsx
